complete code refactor
